http: Extracted netty response handler
